update StateReferenceResolution and make it consistent with others.
