+++ src/integration-test/java/com/commercetools/sync/integration/externalsource/states/StateSyncIT.java
@@ -50,7 +50,7 @@ import static com.commercetools.sync.integration.commons.utils.SphereClientUtils
 import static com.commercetools.sync.integration.commons.utils.StateITUtils.deleteStates;
 import static com.commercetools.sync.integration.commons.utils.StateITUtils.deleteStatesFromTargetAndSource;
 import static com.commercetools.sync.integration.commons.utils.StateITUtils.getStateByKey;
-import static com.commercetools.sync.states.utils.StateTransitionReferenceReplacementUtils.replaceStateTransitionIdsWithKeys;
+import static com.commercetools.sync.states.utils.StateReferenceResolutionUtils.mapToStateDrafts;
 import static com.commercetools.tests.utils.CompletionStageUtil.executeBlocking;
 import static io.sphere.sdk.models.LocalizedString.ofEnglish;
 import static io.sphere.sdk.states.State.referenceOfId;
@@ -562,7 +562,7 @@ class StateSyncIT {
         final State stateB = createStateInSource(stateBDraft);
 
         StateDraft[] draftsWithReplacesKeys =
-            replaceStateTransitionIdsWithKeys(asList(stateB, stateC)).toArray(new StateDraft[2]);
+            mapToStateDrafts(asList(stateB, stateC)).toArray(new StateDraft[2]);
         final StateDraft stateADraft = createStateDraftReferencingStateDrafts(keyA, draftsWithReplacesKeys);
         final List<StateDraft> stateDrafts = asList(stateADraft);
 
@@ -610,7 +610,7 @@ class StateSyncIT {
             .build();
 
         final StateSync stateSync = new StateSync(stateSyncOptions);
-        final List<StateDraft> stateDrafts = replaceStateTransitionIdsWithKeys(Arrays.asList(stateA, stateB, stateC));
+        final List<StateDraft> stateDrafts = mapToStateDrafts(Arrays.asList(stateA, stateB, stateC));
         // test
         final StateSyncStatistics stateSyncStatistics = stateSync
             .sync(stateDrafts)
@@ -661,7 +661,7 @@ class StateSyncIT {
             .build();
 
         final StateSync stateSync = new StateSync(stateSyncOptions);
-        final List<StateDraft> stateDrafts = replaceStateTransitionIdsWithKeys(Arrays.asList(stateA, stateB, stateC));
+        final List<StateDraft> stateDrafts = mapToStateDrafts(Arrays.asList(stateA, stateB, stateC));
         // test
         final StateSyncStatistics stateSyncStatistics = stateSync
             .sync(stateDrafts)
@@ -727,7 +727,7 @@ class StateSyncIT {
             .build();
 
         final StateSync stateSync = new StateSync(stateSyncOptions);
-        final List<StateDraft> stateDrafts = replaceStateTransitionIdsWithKeys(Arrays.asList(stateA, stateB, stateC));
+        final List<StateDraft> stateDrafts = mapToStateDrafts(Arrays.asList(stateA, stateB, stateC));
         // test
         final StateSyncStatistics stateSyncStatistics = stateSync
             .sync(stateDrafts)
@@ -770,7 +770,7 @@ class StateSyncIT {
             .build();
 
         final StateSync stateSync = new StateSync(stateSyncOptions);
-        final List<StateDraft> stateDrafts = replaceStateTransitionIdsWithKeys(Arrays.asList(stateB, stateC));
+        final List<StateDraft> stateDrafts = mapToStateDrafts(Arrays.asList(stateB, stateC));
         // test
         final StateSyncStatistics stateSyncStatistics = stateSync
             .sync(stateDrafts)
@@ -810,7 +810,7 @@ class StateSyncIT {
             .build();
 
         final StateSync stateSync = new StateSync(stateSyncOptions);
-        final List<StateDraft> stateDrafts = replaceStateTransitionIdsWithKeys(Arrays.asList(stateA, stateB, stateC));
+        final List<StateDraft> stateDrafts = mapToStateDrafts(Arrays.asList(stateA, stateB, stateC));
         // test
         final StateSyncStatistics stateSyncStatistics = stateSync
             .sync(stateDrafts)
@@ -853,7 +853,7 @@ class StateSyncIT {
             .build();
 
         final StateSync stateSync = new StateSync(stateSyncOptions);
-        final List<StateDraft> stateDrafts = replaceStateTransitionIdsWithKeys(Arrays.asList(stateA, stateB, stateC));
+        final List<StateDraft> stateDrafts = mapToStateDrafts(Arrays.asList(stateA, stateB, stateC));
         // test
         final StateSyncStatistics stateSyncStatistics = stateSync
             .sync(stateDrafts)
@@ -947,7 +947,7 @@ class StateSyncIT {
                 warningCallBackMessages.add(exception.getMessage()))
             .build();
         final StateSync stateSync = new StateSync(stateSyncOptions);
-        final List<StateDraft> stateDrafts = replaceStateTransitionIdsWithKeys(Arrays.asList(stateA, stateB, stateC));
+        final List<StateDraft> stateDrafts = mapToStateDrafts(Arrays.asList(stateA, stateB, stateC));
         // test
         final StateSyncStatistics stateSyncStatistics = stateSync
             .sync(stateDrafts)
@@ -1016,7 +1016,7 @@ class StateSyncIT {
                 warningCallBackMessages.add(exception.getMessage()))
             .build();
 
-        final List<StateDraft> stateDrafts = replaceStateTransitionIdsWithKeys(Arrays.asList(stateA, stateB, stateC));
+        final List<StateDraft> stateDrafts = mapToStateDrafts(Arrays.asList(stateA, stateB, stateC));
         // test
         final StateSyncStatistics stateSyncStatistics = new StateSync(stateSyncOptions)
             .sync(stateDrafts)
