+++ src/test/java/com/commercetools/sync/states/utils/StateReferenceResolutionUtilsTest.java
@@ -1,8 +1,10 @@
 package com.commercetools.sync.states.utils;
 
+import io.sphere.sdk.expansion.ExpansionPath;
 import io.sphere.sdk.models.Reference;
 import io.sphere.sdk.states.State;
 import io.sphere.sdk.states.StateDraft;
+import io.sphere.sdk.states.queries.StateQuery;
 import org.junit.jupiter.api.Test;
 
 import javax.annotation.Nonnull;
@@ -12,15 +14,16 @@ import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
 
-import static com.commercetools.sync.states.utils.StateTransitionReferenceReplacementUtils.replaceStateTransitionIdsWithKeys;
+import static com.commercetools.sync.states.utils.StateReferenceResolutionUtils.buildStateQuery;
+import static com.commercetools.sync.states.utils.StateReferenceResolutionUtils.mapToStateDrafts;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-class StateReferenceReplacementUtilsTest {
+class StateReferenceResolutionUtilsTest {
 
     @Test
-    void replaceStateReferenceIdsWithKeys_WithAllExpandedReferences_ShouldReturnReferencesWithReplacedKeys() {
+    void mapToStateDrafts_WithAllExpandedReferences_ShouldReturnReferencesWithReplacedKeys() {
         //preparation
         final List<State> mockStates = new ArrayList<>();
         final String stateReferenceKey = "state-key";
@@ -38,7 +41,7 @@ class StateReferenceReplacementUtilsTest {
 
         //test
         final List<StateDraft> referenceReplacedDrafts =
-            replaceStateTransitionIdsWithKeys(mockStates);
+            mapToStateDrafts(mockStates);
 
         //assertion
         referenceReplacedDrafts.forEach(stateDraft -> {
@@ -50,7 +53,7 @@ class StateReferenceReplacementUtilsTest {
     }
 
     @Test
-    void replaceStateReferenceIdsWithKeys_WithAllNonExpandedReferences_ShouldReturnReferencesWithoutReplacedKeys() {
+    void mapToStateDrafts_WithAllNonExpandedReferences_ShouldReturnReferencesWithoutReplacedKeys() {
         //preparation
         final List<State> mockStates = new ArrayList<>();
         final String stateReferenceKey = "state-key";
@@ -67,7 +70,7 @@ class StateReferenceReplacementUtilsTest {
 
         //test
         final List<StateDraft> referenceReplacedDrafts =
-            replaceStateTransitionIdsWithKeys(mockStates);
+            mapToStateDrafts(mockStates);
 
         //assertion
         referenceReplacedDrafts.forEach(stateDraft -> {
@@ -78,7 +81,7 @@ class StateReferenceReplacementUtilsTest {
     }
 
     @Test
-    void replaceStateReferenceIdsWithKeys_WithNullOrEmptyReferences_ShouldNotFail() {
+    void mapToStateDrafts_WithNullOrEmptyReferences_ShouldNotFail() {
         //preparation
         final State mockState1 = mock(State.class);
         when(mockState1.getTransitions()).thenReturn(null);
@@ -88,10 +91,16 @@ class StateReferenceReplacementUtilsTest {
 
         //test
         final List<StateDraft> referenceReplacedDrafts =
-            replaceStateTransitionIdsWithKeys(Arrays.asList(mockState1, mockState2));
+            mapToStateDrafts(Arrays.asList(mockState1, mockState2));
 
         assertThat(referenceReplacedDrafts.get(0).getTransitions()).isEqualTo(Collections.emptySet());
         assertThat(referenceReplacedDrafts.get(1).getTransitions()).isEqualTo(Collections.emptySet());
+    }
+
+    @Test
+    void buildStateQuery_Always_ShouldReturnQueryWithAllNeededReferencesExpanded() {
+        final StateQuery stateQuery = buildStateQuery();
+        assertThat(stateQuery.expansionPaths()).containsExactly(ExpansionPath.of("transitions[*]"));
     }
 
     @Nonnull
